+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 - Sold kilju should not respawn on the junkyard as full anymore
 - Fixed a lag that was fixable by player having walk to the Satsuma
+- More fixed to radiator hose3
 
 ## 3.1.7 (06.04.2021)